backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.12.0
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.3
boto3==1.28.54
botocore==1.31.54
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cohere==4.27
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
diskcache==5.6.1
dnspython==2.4.2
entrypoints==0.4
executing==2.0.0
Faker==19.6.2
fastapi==0.103.2
fastavro==1.8.2
favicon==0.7.0
filelock==3.12.4
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==3.0.0
h11==0.14.0
htbuilder==0.6.2
html2text==2020.1.16
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyterlab-widgets==3.0.9
kaggle==1.5.16
kiwisolver==1.4.5
langchain==0.0.304
langsmith==0.0.41
litellm==0.5.4
llama-index==0.8.41
loguru==0.7.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
more-itertools==10.1.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.24.3
openai==0.27.6
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pinecone-client==2.2.4
pluggy==1.3.0
prompt-toolkit==3.0.38
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.3
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.30.0
rich==13.3.5
s3transfer==0.6.2
safetensors==0.4.0
scipy==1.10.1
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.23.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.21
st-annotated-text==4.0.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
streamlit-aggrid==0.3.4.post3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-javascript==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
text-unidecode==1.3
tiktoken==0.4.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
transformers==4.34.0
trulens-eval==0.14.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
